Rename files, add missing 0.5 SUTs.
