+++ src/modelbench/suts.py
@@ -69,7 +69,9 @@ SUTS_FOR_V_0_5 = [
         "gemma-7b",
         "llama-2-7b-chat",
         "llama-2-13b-chat",
+        "llama-2-70b-chat",
         "mistral-7b",
+        "mixtral-8x-7b",
         "openchat-3_5",
         "qwen-72b",
         "stripedhyena-nous-7b",
